Remove ClientServiceFactory
